+++ src/lib.rs
@@ -143,7 +143,7 @@ pub mod escape {
     pub use crate::escapei::{escape, partial_escape, unescape, unescape_with};
 }
 pub mod events;
-mod name;
+pub mod name;
 mod reader;
 #[cfg(feature = "serialize")]
 pub mod se;